Helper functions to create sparse mask stacks by stamping a template
... and fix a type problem.
Creating the masks as sparse right from the start saves on memory and processing time.

+++ src/libertem/job/masks.py
@@ -137,7 +137,7 @@ class MaskContainer(object):
         # and set the use_sparse property accordingly
 
         if not isinstance(self.mask_factories, Iterable):
-            raw_masks = self.mask_factories()
+            raw_masks = self.mask_factories().astype(self.dtype)
             if self.use_sparse is True:
                 return to_sparse(raw_masks)
             elif self.use_sparse is False:
